# ttms-backend/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ttms影院管理系统",
        "title": "ttms",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/isRePeat": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "判断用户名是否存在",
                "parameters": [
                    {
                        "type": "string",
                        "name": "Username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 20001:用户已存在 20002:用户不存在",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 30001:验证码失效或者有误 20003：密码不能为空",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.LoginRly"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/refreshToken": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "刷新token",
                "parameters": [
                    {
                        "description": "刷新",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RefreshTokenParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 30003:refresh token失效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.RefreshRly"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 20001:用户已存在 ",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.RegisterRly"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/user/findUser": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "根据用户名查找用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "刷新",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FindParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 20002:用户不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.UserInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/user/modifyAvatar": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户修改头像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "newAvatar",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误  20002:用户不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/request.ModifyAvatar"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/user/modifyEmail": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户修改头像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "maxLength": 6,
                        "minLength": 6,
                        "type": "string",
                        "name": "EmailCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "NewEmail",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误  20002:用户不存在 30001:验证码有误",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/user/modifyPassword": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户修改密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "maxLength": 6,
                        "minLength": 6,
                        "type": "string",
                        "name": "EmailCode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "NewPassword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误  20002:用户不存在 30001:验证码有误",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/user/update": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "管理员删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误  20002:用户不存在",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/v1/user/updateInfo": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户修改性别,个性签名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x_token 用户令牌",
                        "name": "x_token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "signature",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误  20002:用户不存在",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "失败时返回空"
                },
                "status_code": {
                    "description": "状态码，0-成功，其他值-失败",
                    "type": "integer"
                },
                "status_msg": {
                    "description": "返回状态描述",
                    "type": "string"
                }
            }
        },
        "reply.LoginRly": {
            "type": "object",
            "properties": {
                "AccessToken": {
                    "type": "string"
                },
                "Payload": {
                    "$ref": "#/definitions/token.Payload"
                },
                "RefreshToken": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_info": {
                    "$ref": "#/definitions/reply.UserInfoReply"
                }
            }
        },
        "reply.RefreshRly": {
            "type": "object",
            "properties": {
                "AccessToken": {
                    "type": "string"
                }
            }
        },
        "reply.RegisterRly": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "reply.UserInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "reply.UserInfoReply": {
            "type": "object",
            "properties": {
                "AvatarURL": {
                    "type": "string"
                },
                "Role": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "request.FindParam": {
            "type": "object",
            "required": [
                "UserName"
            ],
            "properties": {
                "UserName": {
                    "type": "string"
                }
            }
        },
        "request.LoginParam": {
            "type": "object",
            "required": [
                "Email"
            ],
            "properties": {
                "Email": {
                    "type": "string"
                },
                "EmailCode": {
                    "type": "string"
                },
                "LoginType": {
                    "description": "1表示验证码登录，2表示密码登录",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "Password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "request.ModifyAvatar": {
            "type": "object",
            "required": [
                "newAvatar"
            ],
            "properties": {
                "newAvatar": {
                    "type": "string"
                }
            }
        },
        "request.RefreshTokenParam": {
            "type": "object",
            "required": [
                "AccessToken",
                "RefreshToken"
            ],
            "properties": {
                "AccessToken": {
                    "type": "string"
                },
                "RefreshToken": {
                    "type": "string"
                }
            }
        },
        "request.RegisterParam": {
            "type": "object",
            "required": [
                "Email",
                "EmailCode",
                "Password",
                "UserName"
            ],
            "properties": {
                "Avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "Email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "EmailCode": {
                    "description": "邮箱验证码，长度为6",
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6
                },
                "Gender": {
                    "description": "性别",
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ]
                },
                "Password": {
                    "description": "密码长度介于3和12之间",
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 3
                },
                "Signature": {
                    "description": "个性签名",
                    "type": "string"
                },
                "UserName": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "token.Payload": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "可以是用户或者是账户",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "expired-at": {
                    "type": "string"
                },
                "id": {
                    "description": "用于管理每个JWT",
                    "type": "string"
                },
                "issued-at": {
                    "description": "创建时间用于检验",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}